Added Alessandra's new models
Today Alessandra completed the police car, the coin, heart, magnent, and x2 star. I added them to the build and put their textures on them. I still need to make prefabs of them.  I gave some feedback to alessandra to fix some of the issues that I saw with them.