Score Manager & Speed Manager
I changed the way that the score multiplier works. I've now linked it to the length of time the player has been alive in the level up to x10. With a visual indicator of how long until it increases.

I also created a speed manager that speeds the player up over time and increases the difficulty of the game.